+++ src/screens/courses-list/dialogs/join-course/JoinCourseDialog.tsx
