+++ src/screens/courses-list/dialogs/join-course/JoinCourseForm.tsx
