+++ src/screens/index.ts
@@ -7,18 +7,18 @@ export { RegisterAdminForm } from "./session/register-admin/Form";
 export { RegisterTeacherForm } from "./session/register-teacher/Form";
 
 // Admin views
-export { AdminsView } from "./admins-view/AdminsView";
+export { AdminsView } from "./admins-list/AdminsView";
 
 // Course Home
 export { CoursePageLayout } from "./course-page/CoursePageLayout";
-export { CoursesHome } from "./courses-home/CoursesHome";
+export { CoursesHome } from "./courses-list/CoursesHome";
 
 // Course details views
 export { CourseLaboratories } from "./course-page/laboratories/CourseLaboratories";
 export { CourseParticipants } from "./course-page/participants/CourseParticipants";
 
 // Rubrics views
-export { RubricsHome } from "./rubrics-home/RubricsHome";
+export { RubricsHome } from "./rubrics-list/RubricsHome";
 export { EditRubricView } from "./edit-rubric/EditRubricView";
 
 // Misc views
