+++ src/screens/rubrics-list/components/RubricsTable.tsx
